Works without pyproject.toml
I think that's fine though...
