+++ MANIFEST.in
@@ -1,5 +1,5 @@
 global-include README.md VERSION LICENSE
-global-include setup.py
+global-include setup.py build_helper.py
 graft src
 prune test
 prune build
