+++ setup.py
@@ -1,7 +1,14 @@
 from os.path import join as opj
 
 from setuptools import setup, Extension, find_packages
-from setuptools.command.bdist_egg import bdist_egg as _bdist_egg
+
+from build_helper import zf_cythonize, build_wavefront, no_egg
+
+try:
+    from sage_setup.command.sage_build_cython import sage_build_cython
+    SAGE_INSTALLED = True
+except ImportError:
+    SAGE_INSTALLED = False
 
 classifiers = [
     "Development Status :: 4 - Beta",
@@ -15,50 +22,6 @@ classifiers = [
     "Topic :: Scientific/Engineering :: Mathematics",
     "Topic :: Software Development :: Libraries :: Python Modules",
 ]
-
-try:
-    from sage_setup.command.sage_build_cython import sage_build_cython
-    from sage_setup.command.sage_build_ext import sage_build_ext as _build_ext
-
-    SAGE_INSTALLED = True
-except ImportError:
-    from setuptools.command.build_ext import build_ext as _build_ext
-    SAGE_INSTALLED = False
-
-class no_egg(_bdist_egg):
-    def run(self):
-        from distutils.errors import DistutilsOptionError
-        raise DistutilsOptionError("Honestly just copying https://github.com/sagemath/cysignals/blob/c901dc9217de735c67ca5daf3dff6276813a05b5/setup.py#L186-L193")
-
-class zf_cythonize(_build_ext):
-    base_directives = dict(
-         binding=False,
-         language_level=3,
-    )
-    def finalize_options(self):
-        dist = self.distribution
-        ext_modules = dist.ext_modules
-        if ext_modules:
-            dist.ext_modules[:] = self.cythonize(ext_modules)
-        super().finalize_options()
-
-    def cythonize(self, extensions):
-        # Run Cython with -Werror on continuous integration services
-        # with Python 3.6 or later
-        from Cython.Compiler import Options
-        Options.warning_errors = False
-
-        compiler_directives = dict(**self.base_directives)
-        from Cython.Build.Dependencies import cythonize
-        return cythonize(extensions,
-                         compiler_directives=compiler_directives)
-
-class build_wavefront(zf_cythonize):
-    def initialize_options(self):
-        super().initialize_options()
-        ext_name = "zeroforcing.test.verifiability.wavefront"
-        self.distribution.packages = [ext_name]
-        self.distribution.ext_modules = [Extension(ext_name, sources=[opj("test", "verifiability", "wavefront.pyx")])]
 
 with open("VERSION") as f:
     VERSION = f.read().strip()
